# backend/openapi.yaml
openapi: 3.0.3
info:
  title: ITFY E-Voting Platform API
  description: |
    # ITFY E-Voting Platform Backend API
    
    Welcome to the comprehensive API documentation for the ITFY E-Voting platform.
    This platform enables event organizers to create and manage voting events with 
    candidates, categories, and secure payment-based voting.
    
    ## Features
    - **Authentication & Authorization**: Secure JWT-based authentication with role-based access control
    - **Event Management**: Create, manage, and publish voting events
    - **Category Management**: Organize candidates into voting categories
    - **Candidate Management**: Register and manage candidates with profiles
    - **Voting System**: Purchase vote bundles and cast votes securely
    - **Payment Integration**: Paystack payment gateway integration
    - **Forms & Submissions**: Dynamic nomination forms with approval workflows
    - **Analytics**: Comprehensive voting analytics and reporting
    
    ## Authentication
    Most endpoints require authentication via JWT tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_token>
    ```
    
    ## User Roles
    - **super_admin**: Full system access
    - **admin**: Administrative access to events and users
    - **organiser**: Can create and manage events
    - **moderator**: Can view and moderate content
    
    ## Rate Limiting
    API requests are rate-limited to prevent abuse:
...
    
    ## Pagination
...
    
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
    CandidateAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated candidates
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "An error occurred"
        error:
          type: string
          example: "Detailed error message"
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation successful"
        data:
          type: object
    PaginationMeta:
      type: object
      properties:
        currentPage:
          type: integer
          example: 1
        totalPages:
          type: integer
          example: 10
        totalItems:
          type: integer
          example: 95
        itemsPerPage:
          type: integer
          example: 10
        hasNextPage:
          type: boolean
          example: true
        hasPrevPage:
          type: boolean
          example: false
    ObjectId:
      type: string
      pattern: '^[a-fA-F0-9]{24}$'
      example: "507f1f77bcf86cd799439011"

    # ==================
    # User Schemas
    # ==================
...
    Vote:
      type: object
      properties:
        _id:
          $ref: '#/components/schemas/ObjectId'
        candidate:
          $ref: '#/components/schemas/ObjectId'
        category:
          $ref: '#/components/schemas/ObjectId'
        event:
          $ref: '#/components/schemas/ObjectId'
        payment:
          $ref: '#/components/schemas/ObjectId'
        vote_code:
          type: string
          description: Vote code used
        status:
          type: string
          enum: [active, refunded]
        ip_hash:
          type: string
          description: Hashed IP address
        user_agent:
          type: string
        cast_at:
          type: string
          format: date-time
        refunded_at:
          type: string
          format: date-time
          nullable: true
        refund_reason:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CastVoteRequest:
      type: object
      required:
        - vote_code
        - candidate
        - category
      properties:
        vote_code:
          type: string
          description: Vote code from payment
        candidate:
          $ref: '#/components/schemas/ObjectId'
        category:
          $ref: '#/components/schemas/ObjectId'
    VoteEligibilityRequest:
      type: object
      required:
        - vote_code
        - category
      properties:
        vote_code:
          type: string
        category:
          $ref: '#/components/schemas/ObjectId'

    # ==================
    # Bundle Schemas
    # ==================
...
  responses:
    UnauthorizedError:
      description: Authentication token is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "Unauthorized"
            error: "You are not authorized to perform this action"
    ForbiddenError:
      description: User does not have permission to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "Forbidden"
            error: "Access denied. You don't have permission for this resource"
    NotFoundError:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "Not Found"
            error: "The requested resource was not found"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "Validation Error"
            error: "Some required fields are missing or invalid"
...
  parameters:
    PageParam:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    LimitParam:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: Number of items per page
...
    IdParam:
      in: path
      name: id
      required: true
      schema:
        $ref: '#/components/schemas/ObjectId'
      description: Resource ID
...
paths:
...
  /events/{id}/results:
    get:
      tags:
        - Events
      summary: Get event results
      description: Retrieve voting results for an event
      operationId: getEventResults
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Results retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '403':
          description: Results not yet published
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /categories/{id}/results:
    get:
      tags:
        - Categories
      summary: Get category results
      description: Retrieve voting results for a category (visibility based on settings)
      operationId: getCategoryResults
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Results retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '403':
          description: Results not visible
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /votes:
    post:
      tags:
        - Votes
      summary: Cast a vote
      description: Cast a vote using a vote code
      operationId: castVote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CastVoteRequest'
      responses:
        '201':
          description: Vote cast successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Vote'
        '400':
          $ref: '#/components/responses/ValidationError'

    get:
      tags:
        - Votes
      summary: List all votes
      description: Retrieve paginated list of votes (Admin only)
      operationId: listVotes
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - in: query
          name: event
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: candidate
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [active, refunded]
      responses:
        '200':
          description: Votes retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Vote'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /votes/validate-eligibility:
    post:
      tags:
        - Votes
      summary: Validate vote eligibility
      description: Check if vote code is eligible to vote in a category
      operationId: validateVoteEligibility
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoteEligibilityRequest'
      responses:
        '200':
          description: Eligibility check result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      eligible:
                        type: boolean
                      reason:
                        type: string
                      votes_remaining:
                        type: integer
...
  /votes/history/{code}:
    get:
      tags:
        - Votes
      summary: Get voter history
      description: Retrieve voting history for a vote code
      operationId: getVoterHistory
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
      responses:
        '200':
          description: History retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
...
  /votes/count/candidate/{candidateId}:
    get:
      tags:
        - Votes
      summary: Get candidate vote count
      description: Retrieve vote count for a candidate
      operationId: getVoteCountByCandidate
      parameters:
        - in: path
          name: candidateId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Count retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      count:
                        type: integer
  /votes/count/category/{categoryId}:
    get:
      tags:
        - Votes
      summary: Get category vote count
      description: Retrieve total votes in a category
      operationId: getVoteCountByCategory
      parameters:
        - in: path
          name: categoryId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Count retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      count:
                        type: integer
  /votes/count/event/{eventId}:
    get:
      tags:
        - Votes
      summary: Get event vote count
      description: Retrieve total votes in an event
      operationId: getVoteCountByEvent
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Count retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      count:
                        type: integer
  /votes/trends/{eventId}:
    get:
      tags:
        - Votes
      summary: Get voting trends
      description: Retrieve voting trends for an event (Admin only)
      operationId: getVotingTrends
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Trends retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /votes/distribution/{eventId}:
    get:
      tags:
        - Votes
      summary: Get vote distribution
      description: Retrieve vote distribution by category (Admin only)
      operationId: getVoteDistribution
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Distribution retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /votes/top-candidates/{eventId}:
    get:
      tags:
        - Votes
      summary: Get top candidates
      description: Retrieve top candidates for an event (Admin only)
      operationId: getTopCandidates
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Top candidates retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /votes/stats/{eventId}:
    get:
      tags:
        - Votes
      summary: Get voting statistics
      description: Retrieve voting statistics for an event (Admin only)
      operationId: getVotingStats
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Statistics retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
  /votes/candidate/{candidateId}:
    get:
      tags:
        - Votes
      summary: Get votes by candidate
      description: Retrieve votes for a candidate (Admin only)
      operationId: getVotesByCandidate
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: candidateId
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Votes retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Vote'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /votes/category/{categoryId}:
    get:
      tags:
        - Votes
      summary: Get votes by category
      description: Retrieve votes for a category (Admin only)
      operationId: getVotesByCategory
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: categoryId
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Votes retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Vote'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /votes/event/{eventId}:
    get:
      tags:
        - Votes
      summary: Get votes by event
      description: Retrieve votes for an event (Admin only)
      operationId: getVotesByEvent
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Votes retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Vote'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /votes/{id}:
    get:
      tags:
        - Votes
      summary: Get vote by ID
      description: Retrieve vote details (Admin only)
      operationId: getVoteById
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Vote retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Vote'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /votes/{id}/refund:
    post:
      tags:
        - Votes
      summary: Refund a vote
      description: Refund a vote (Admin only)
      operationId: refundVote
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Reason for refund
      responses:
        '200':
          description: Vote refunded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # ==================
  # Bundle Endpoints
  # ==================
...